devstack/acceptance test : rename FollowSource to FollowL2
